# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TukAPI
  version: 1.0.0
  description: >
    OpenAPI draft for the Sri Lanka Police tuk-tuk tracking and movement logging API.
    This document covers authentication, protected master-data management, device ping
    ingestion, and live or historical location lookup endpoints.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /stations:
    get:
      tags: [Stations]
      summary: List police stations
      operationId: listStations
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/ProvinceId'
        - $ref: '#/components/parameters/DistrictId'
        - $ref: '#/components/parameters/IsActive'
        - $ref: '#/components/parameters/StationSortBy'
        - $ref: '#/components/parameters/SortOrderAsc'
      responses:
        '200':
          description: Police station list returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Stations]
      summary: Create a police station
      operationId: createStation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStationRequest'
      responses:
        '201':
          description: Police station created successfully
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stations/{id}:
    get:
      tags: [Stations]
      summary: Get a single police station
      operationId: getStationById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Police station returned successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationSingleResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Stations]
      summary: Partially update a police station
      operationId: updateStation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStationRequest'
      responses:
        '200':
          description: Police station updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drivers:
    get:
      tags: [Drivers]
      summary: List drivers
      operationId: listDrivers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/ProvinceId'
        - $ref: '#/components/parameters/DistrictId'
        - $ref: '#/components/parameters/StationId'
        - $ref: '#/components/parameters/IsActive'
        - $ref: '#/components/parameters/LicenseNumber'
        - $ref: '#/components/parameters/NationalId'
        - $ref: '#/components/parameters/DriverSortBy'
        - $ref: '#/components/parameters/SortOrderDesc'
      responses:
        '200':
          description: Driver list returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Drivers]
      summary: Create a driver
      operationId: createDriver
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDriverRequest'
      responses:
        '201':
          description: Driver created successfully
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /drivers/{id}:
    get:
      tags: [Drivers]
      summary: Get a single driver
      operationId: getDriverById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Driver returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverSingleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Drivers]
      summary: Partially update a driver
      operationId: updateDriver
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDriverRequest'
      responses:
        '200':
          description: Driver updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /devices:
    get:
      tags: [Devices]
      summary: List tracking devices
      operationId: listDevices
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/TukTukId'
        - $ref: '#/components/parameters/DeviceStatus'
        - $ref: '#/components/parameters/SerialNumber'
        - $ref: '#/components/parameters/DeviceSortBy'
        - $ref: '#/components/parameters/SortOrderDesc'
      responses:
        '200':
          description: Device list returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Devices]
      summary: Create a tracking device
      operationId: createDevice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceRequest'
      responses:
        '201':
          description: Device created successfully
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /devices/{id}:
    get:
      tags: [Devices]
      summary: Get a single tracking device
      operationId: getDeviceById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Device returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceSingleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Devices]
      summary: Partially update a tracking device
      operationId: updateDevice
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceRequest'
      responses:
        '200':
          description: Device updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /locations/history:
    get:
      tags: [Locations]
      summary: Query historical location pings
      operationId: getLocationHistory
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/ProvinceId'
        - $ref: '#/components/parameters/DistrictId'
        - $ref: '#/components/parameters/StationId'
        - $ref: '#/components/parameters/TukTukId'
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/FromTime'
        - $ref: '#/components/parameters/ToTime'
        - $ref: '#/components/parameters/IgnitionOn'
        - $ref: '#/components/parameters/LocationHistorySortBy'
        - $ref: '#/components/parameters/SortOrderDesc'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Location history returned successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationPingListResponse'
              example:
                success: true
                message: Movement history fetched successfully.
                data:
                  - id: cm9x7a1bc0010abc123456789
                    tukTukId: cm9x7a1bc0006abc123456789
                    deviceId: cm9x7a1bc0005abc123456789
                    provinceId: cm9x7a1bc0001abc123456789
                    districtId: cm9x7a1bc0002abc123456789
                    stationId: cm9x7a1bc0003abc123456789
                    latitude: 6.927079
                    longitude: 79.861244
                    speedKmh: 34.6
                    heading: 176.4
                    accuracyM: 5.2
                    ignitionOn: true
                    recordedAt: '2026-04-29T06:25:12.000Z'
                    receivedAt: '2026-04-29T06:25:13.000Z'
                    createdAt: '2026-04-29T06:25:13.000Z'
                meta:
                  page: 1
                  limit: 20
                  totalItems: 1
                  totalPages: 1
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated law-enforcement users.
    DeviceTokenAuth:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Device token used for telemetry ingestion routes.
  headers:
    ETag:
      description: Weak entity tag for conditional GET support.
      schema:
        type: string
      example: W/"S2JkQm5GQ1hPVEp5R0FxQ0pmVQ"
    LastModified:
      description: Last modification timestamp for conditional GET support.
      schema:
        type: string
        format: date-time
      example: '2026-04-29T06:25:13.000Z'
    CacheControl:
      description: Cache directive for the returned representation.
      schema:
        type: string
      example: private, max-age=30, must-revalidate
    Location:
      description: Relative URI of the created resource.
      schema:
        type: string
      example: /api/v1/tuk-tuks/cm9x7a1bc0006abc123456789
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier.
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of records per page.
    Search:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search term.
    ProvinceId:
      name: provinceId
      in: query
      schema:
        type: string
      description: Filter by province ID.
...
    DistrictId:
      name: districtId
      in: query
      schema:
        type: string
      description: Filter by district ID.
    StationId:
      name: stationId
      in: query
      schema:
        type: string
      description: Filter by station ID.
...
    DeviceId:
      name: deviceId
      in: query
      schema:
        type: string
      description: Filter by device ID.
    TukTukId:
      name: tukTukId
      in: query
      schema:
        type: string
      description: Filter by tuk-tuk ID.
    IsActive:
      name: isActive
      in: query
      schema:
        type: boolean
      description: Filter by active status.
    LicenseNumber:
      name: licenseNumber
      in: query
      schema:
        type: string
      description: Filter by exact driver license number.
    NationalId:
      name: nationalId
      in: query
      schema:
        type: string
      description: Filter by exact driver national ID.
...
    SerialNumber:
      name: serialNumber
      in: query
      schema:
        type: string
      description: Filter by exact device serial number.
...
    DeviceStatus:
      name: status
      in: query
      schema:
        $ref: '#/components/schemas/DeviceStatus'
      description: Filter by device status.
...
    StationSortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [createdAt, updatedAt, name, code]
        default: name
      description: Police station sort field.
    DriverSortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [createdAt, updatedAt, fullName, licenseNumber]
        default: createdAt
      description: Driver sort field.
...
    DeviceSortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [createdAt, updatedAt, serialNumber, status]
        default: createdAt
      description: Device sort field.
...
    LocationHistorySortBy:
      name: sortBy
      in: query
      schema:
        type: string
        enum: [recordedAt, receivedAt, createdAt]
        default: recordedAt
      description: History sort field.
    SortOrderAsc:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
      description: Sort direction.
    SortOrderDesc:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort direction.
...
    FromTime:
      name: from
      in: query
      schema:
        type: string
        format: date-time
      description: Inclusive start of the requested time range.
    ToTime:
      name: to
      in: query
      schema:
        type: string
        format: date-time
      description: Inclusive end of the requested time range.
    IgnitionOn:
      name: ignitionOn
      in: query
      schema:
        type: boolean
      description: Filter by ignition state.
    IfNoneMatch:
      name: If-None-Match
      in: header
      schema:
        type: string
      description: Conditional GET validator based on the last returned ETag.
    IfModifiedSince:
      name: If-Modified-Since
      in: header
      schema:
        type: string
        format: date-time
      description: Conditional GET validator based on resource modification time.
  responses:
    NotModified:
      description: Resource has not changed since the supplied cache validator.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        Last-Modified:
          $ref: '#/components/headers/LastModified'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
    BadRequest:
      description: Request failed validation or query parsing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Validation failed.
            errors:
              - path: query.limit
                message: Number must be less than or equal to 100.
    Unauthorized:
      description: Authentication failed or credentials are missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication token is required.
    Forbidden:
      description: Authenticated caller does not have access to this resource or scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: You are not authorized to access this resource.
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Requested resource was not found.
    Conflict:
      description: Request would violate a unique constraint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: A unique constraint would be violated by this request.
    UnprocessableEntity:
      description: Request is valid JSON but violates a business rule.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Tracking device is already assigned to another tuk-tuk.
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Too many requests, please try again later.
    InternalServerError:
      description: Unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Internal server error.
  schemas:
...
    PoliceStation:
      type: object
      properties:
        id:
          type: string
        provinceId:
          type: string
        districtId:
          type: string
        name:
          type: string
        code:
          type: string
        address:
          type: string
        contactNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        province:
          $ref: '#/components/schemas/ProvinceSummary'
        district:
          $ref: '#/components/schemas/DistrictSummary'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - provinceId
        - districtId
        - name
        - code
        - address
        - isActive
        - province
        - district
        - createdAt
        - updatedAt
...
    StationSingleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Police station fetched successfully.
        data:
          $ref: '#/components/schemas/PoliceStation'
      required: [success, message, data]
    StationListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Police stations fetched successfully.
        data:
          type: array
          items:
            $ref: '#/components/schemas/PoliceStation'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [success, message, data, meta]
    DriverSingleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Driver fetched successfully.
        data:
          $ref: '#/components/schemas/Driver'
      required: [success, message, data]
    DriverListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Drivers fetched successfully.
        data:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [success, message, data, meta]
...
    DeviceSingleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Device fetched successfully.
        data:
          $ref: '#/components/schemas/TrackingDevice'
      required: [success, message, data]
    DeviceListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Devices fetched successfully.
        data:
          type: array
          items:
            $ref: '#/components/schemas/TrackingDevice'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [success, message, data, meta]
...
    LocationPingListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Location history fetched successfully.
        data:
          type: array
          items:
            $ref: '#/components/schemas/LocationPing'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [success, message, data, meta]
...
    CreateStationRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        provinceId:
          type: string
        districtId:
          type: string
        isActive:
          type: boolean
          default: true
      required: [name, code, address, provinceId, districtId]
    UpdateStationRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        provinceId:
          type: string
        districtId:
          type: string
        isActive:
          type: boolean
    CreateDriverRequest:
      type: object
      properties:
        fullName:
          type: string
        nationalId:
          type: string
        licenseNumber:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
        stationId:
          type: string
        isActive:
          type: boolean
          default: true
      required: [fullName, nationalId, licenseNumber, stationId]
    UpdateDriverRequest:
      type: object
      properties:
        fullName:
          type: string
        nationalId:
          type: string
        licenseNumber:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
        stationId:
          type: string
        isActive:
          type: boolean
...
    CreateDeviceRequest:
      type: object
      properties:
        serialNumber:
          type: string
        imei:
          type: string
        simNumber:
          type: string
        firmwareVersion:
          type: string
        status:
          $ref: '#/components/schemas/DeviceStatus'
        tukTukId:
          type: string
          nullable: true
      required: [serialNumber]
    UpdateDeviceRequest:
      type: object
      properties:
        serialNumber:
          type: string
        imei:
          type: string
        simNumber:
          type: string
        firmwareVersion:
          type: string
        status:
          $ref: '#/components/schemas/DeviceStatus'
        tukTukId:
          type: string
          nullable: true
...